Replace current with new implementation
