+++ mbed_build/_internal/find_files.py
@@ -2,72 +2,78 @@
 # Copyright (C) 2020 Arm Mbed. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
-"""File scanner."""
+"""Find files in MbedOS program directory."""
 from pathlib import Path
-from fnmatch import fnmatch
-from typing import Iterable, List
+import fnmatch
+from typing import Callable, Iterable, Optional, List, Tuple
+
 from mbed_targets import get_build_attributes_by_board_type
 
 
-def find_files(file_name: str, directory: Path) -> Iterable[Path]:
-    """Recursively find files matching name under given directory."""
-    return directory.rglob(file_name)
+def find_files(filename: str, directory: Path, filters: Optional[List[Callable]] = None) -> List[Path]:
+    """Recursively find files by name under a given directory.
 
+    This function automatically applies rules from .mbedignore files found during traversal.
 
-_LEGACY_DIRS = [
-    "TESTS",
-    "TEST_APPS",
-]
-
-
-def exclude_legacy_directories(paths: Iterable[Path]) -> List[Path]:
-    """Filter out given paths using legacy directory list.
-
-    WARNING: This should be removed when mbed-os is cleaned up
-    """
-    return [path for path in paths if not any(legacy_dir in str(path) for legacy_dir in _LEGACY_DIRS)]
-
-
-def exclude_using_mbedignore(directory: Path, paths: Iterable[Path]) -> List[Path]:
-    """Filter out given paths based on rules found in .mbedignore files.
-
-    Patterns in .mbedignore use unix shell-style wildcards (fnmatch). It means
-    that functionality, although similar is different to that found in
-    .gitignore and friends.
+    It is important to realise that applied filters are "greedy". The moment a directory is filtered out,
+    its children won't be traversed.
 
     Args:
-        mbedignore_file: Path to .mbedignore.
-        paths: Paths to filter.
-
-    Returns:
-        List of paths.
+        filename: Name of the file to look for.
+        directory: Location where search starts.
+        filters: Optional list of exclude filters to apply.
     """
-    result = list(paths)
-    mbedignore_paths = find_files(".mbedignore", directory)
-    for mbedignore_path in mbedignore_paths:
-        patterns = _build_mbedignore_patterns(mbedignore_path)
-        result = [path for path in result if not _matches_mbedignore_patterns(path, patterns)]
+    if filters is None:
+        filters = []
+
+    result: List[Path] = []
+
+    # Directories and files to process
+    children = list(directory.iterdir())
+
+    # If .mbedignore is one of the children, we need to add it to filter list,
+    # as it might contain rules for currently processed directory, as well as its descendats.
+    mbedignore = Path(directory, ".mbedignore")
+    if mbedignore in children:
+        filters = filters + [MbedignoreFilter.from_file(mbedignore)]
+
+    # Remove files and directories that don't match current set of filters
+    filtered_children = (child for child in children if all(f(child) for f in filters))
+
+    for child in filtered_children:
+        if child.is_dir():
+            # If processed child is a directory, recurse with current set of filters
+            result += find_files(filename, child, filters)
+
+        if child.is_file() and child.name == filename:
+            # We've got a match
+            result.append(child)
+
     return result
 
 
-def exclude_using_target_labels(mbed_program_directory: Path, board_type: str, paths: Iterable[Path]) -> List[Path]:
-    """Filter out given paths using target labels retrieved from mbed-targets.
+class BoardLabelFilter:
+    """Filter out given paths using path labelling rules specific to a board."""
 
-    Args:
-        mbed_program_directory: Path to mbed program, used by mbed-targets to retrieve target build attributes.
-        board_type: Board type, used by mbed-targets to retrieve target build attributes.
-        paths: Paths to filter.
-    """
-    build_attributes = get_build_attributes_by_board_type(board_type, mbed_program_directory)
-    result = list(paths)
-    result = exclude_using_labels("TARGET", build_attributes.labels, result)
-    result = exclude_using_labels("FEATURE", build_attributes.features, result)
-    result = exclude_using_labels("COMPONENT", build_attributes.components, result)
-    return result
+    def __init__(self, board_type: str, mbed_program_directory: Path):
+        """Initialise filter attributes.
+
+        Allowed label data will be retrieved from `mbed-targets`.
+        """
+        build_attributes = get_build_attributes_by_board_type(board_type, mbed_program_directory)
+        self._label_filters = [
+            LabelFilter("TARGET", build_attributes.labels),
+            LabelFilter("FEATURE", build_attributes.features),
+            LabelFilter("COMPONENT", build_attributes.components),
+        ]
+
+    def __call__(self, path: Path) -> bool:
+        """Return True if given path contains only allowed labels - should not be filtered out."""
+        return all(f(path) for f in self._label_filters)
 
 
-def exclude_using_labels(label_type: str, allowed_label_values: Iterable[str], paths: Iterable[Path]) -> List[Path]:
-    """Filter out given path objects using path labelling rules.
+class LabelFilter:
+    """Filter out given paths using path labelling rules.
 
     If a path is labelled with given type, but contains label value which is
     not allowed, it will be filtered out.
@@ -79,38 +85,53 @@ def exclude_using_labels(label_type: str, allowed_label_values: Iterable[str], p
     - "/path/FEATURE_FOO/somefile.txt" will not be filtered out
     - "/path/FEATURE_BAZ/somefile.txt" will be filtered out
     - "/path/FEATURE_FOO/FEATURE_BAR/somefile.txt" will be filtered out
-
-    Args:
-        label_type: Type of label.
-        allowed_label_values: Labels which are allowed for given type.
-        paths: Paths to filter.
     """
-    result = []
-    allowed_labels = set(f"{label_type}_{label_value}" for label_value in allowed_label_values)
-    for path in paths:
-        labels = set(part for part in path.parts if label_type in part)
-        if labels.issubset(allowed_labels):
-            result.append(path)
-    return result
+
+    def __init__(self, label_type: str, allowed_label_values: Iterable[str]):
+        """Initialise the filter attributes.
+
+        Args:
+            label_type: Type of the label to filter with. In filtered paths, it prefixes the value.
+            allowed_label_values: Values which are allowed for the given label type.
+        """
+        self._label_type = label_type
+        self._allowed_labels = set(f"{label_type}_{label_value}" for label_value in allowed_label_values)
+
+    def __call__(self, path: Path) -> bool:
+        """Return True if given path contains only allowed labels - should not be filtered out."""
+        labels = set(part for part in path.parts if self._label_type in part)
+        return labels.issubset(self._allowed_labels)
 
 
-def _build_mbedignore_patterns(mbedignore_path: Path) -> Iterable[str]:
-    """Return patterns extracted from .mbedignore file.
+class MbedignoreFilter:
+    """Filter out given paths based on rules found in .mbedignore files.
 
-    Filters out commented out and empty lines.
-    Prefixes each rule with the directory location of the .mbedignore file.
-
-    Args:
-        mbedignore_path: Path to .mbedignore.
+    Patterns in .mbedignore use unix shell-style wildcards (fnmatch). It means
+    that functionality, although similar is different to that found in
+    .gitignore and friends.
     """
-    lines = mbedignore_path.read_text().splitlines()
-    pattern_lines = (line for line in lines if line.strip() and not line.startswith("#"))
-    ignore_root = mbedignore_path.parent
-    patterns = tuple(str(ignore_root.joinpath(pattern)) for pattern in pattern_lines)
-    return patterns
 
+    def __init__(self, patterns: Tuple[str, ...]):
+        """Initialise the filter attributes.
 
-def _matches_mbedignore_patterns(path: Path, patterns: Iterable[str]) -> bool:
-    """Check if given path matches one of the .mbedignore patterns."""
-    stringified = str(path)
-    return any(fnmatch(stringified, pattern) for pattern in patterns)
+        Args:
+            patterns: List of patterns from .mbedignore to filter against.
+        """
+        self._patterns = patterns
+
+    def __call__(self, path: Path) -> bool:
+        """Return True if given path doesn't match .mbedignore patterns - should not be filtered out."""
+        stringified = str(path)
+        return not any(fnmatch.fnmatch(stringified, pattern) for pattern in self._patterns)
+
+    @classmethod
+    def from_file(cls, mbedignore_path: Path) -> "MbedignoreFilter":
+        """Return new instance with patterns read from .mbedignore file.
+
+        Constructed patterns are rooted in the directory of .mbedignore file.
+        """
+        lines = mbedignore_path.read_text().splitlines()
+        pattern_lines = (line for line in lines if line.strip() and not line.startswith("#"))
+        ignore_root = mbedignore_path.parent
+        patterns = tuple(str(ignore_root.joinpath(pattern)) for pattern in pattern_lines)
+        return cls(patterns)
