--- tests/_internal/test_new_find_files.py
@@ -1,168 +1,0 @@
-#
-# Copyright (C) 2020 Arm Mbed. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-import contextlib
-from pathlib import Path
-from tempfile import TemporaryDirectory
-from unittest import TestCase, mock
-from typing import Iterable
-
-from mbed_targets import MbedTargetBuildAttributes
-
-from mbed_build._internal.new_find_files import find_files, MbedignoreFilter, LabelFilter, BoardLabelFilter
-
-
-@contextlib.contextmanager
-def create_files(files: Iterable[Path]):
-    with TemporaryDirectory() as temp_directory:
-        temp_directory = Path(temp_directory)
-        for file in files:
-            file_path = temp_directory / file
-            file_directory = file_path.parent
-            if not file_directory.exists():
-                file_directory.mkdir(parents=True)
-            file_path.touch()
-        yield Path(temp_directory)
-
-
-class TestListFiles(TestCase):
-    def test_finds_files_by_name(self):
-        matching_paths = [
-            Path("file.txt"),
-            Path("sub_directory", "file.txt"),
-        ]
-        excluded_paths = [
-            Path("not_interested.txt"),
-            Path("sub_directory", "not_interested.txt"),
-        ]
-
-        with create_files(matching_paths + excluded_paths) as directory:
-            subject = find_files("file.txt", directory)
-
-        self.assertEqual(len(subject), len(matching_paths))
-        for path in matching_paths:
-            self.assertIn(Path(directory, path), subject)
-
-    def test_respects_mbedignore(self):
-        matching_paths = [
-            Path("file.txt"),
-        ]
-        excluded_paths = [
-            Path("foo", "file.txt"),
-            Path("bar", "file.txt"),
-        ]
-        with create_files(matching_paths + excluded_paths) as directory:
-            Path(directory, ".mbedignore").write_text("foo/*")
-            Path(directory, "bar", ".mbedignore").write_text("*")
-
-            subject = find_files("file.txt", directory)
-
-        self.assertEqual(len(subject), len(matching_paths))
-        for path in matching_paths:
-            self.assertIn(Path(directory, path), subject)
-
-    def test_respects_given_filters(self):
-        matching_paths = [
-            Path("foo", "file.txt"),
-            Path("hello", "world", "file.txt"),
-        ]
-        excluded_paths = [
-            Path("bar", "file.txt"),
-            Path("foo", "foo-bar", "file.txt"),
-        ]
-
-        def my_filter(path):
-            return "bar" not in str(path)
-
-        with create_files(matching_paths + excluded_paths) as directory:
-            subject = find_files("file.txt", directory, [my_filter])
-
-        self.assertEqual(len(subject), len(matching_paths))
-        for path in matching_paths:
-            self.assertIn(Path(directory, path), subject)
-
-
-@mock.patch("mbed_build._internal.new_find_files.get_build_attributes_by_board_type", autospec=True)
-class TestBoardLabelFilter(TestCase):
-    def test_matches_paths_not_following_board_label_rules(self, get_build_attributes_by_board_type):
-        build_attributes = mock.Mock(
-            MbedTargetBuildAttributes, labels={"T"}, features={"F"}, components={"C"}
-        )  # Zero interface safety here, dataclasses don't support spec_set
-        get_build_attributes_by_board_type.return_value = build_attributes
-        mbed_program_directory = Path("some-program")
-        board_type = "K64F"
-
-        subject = BoardLabelFilter(board_type, mbed_program_directory)
-
-        self.assertFalse(subject(Path("TARGET_X")))
-        self.assertFalse(subject(Path("TARGET_T", "FEATURE_X")))
-        self.assertFalse(subject(Path("TARGET_T", "FEATURE_F", "COMPONENT_X")))
-
-        get_build_attributes_by_board_type.assert_called_with(board_type, mbed_program_directory)
-
-    def test_does_not_match_paths_following_board_label_rules(self, get_build_attributes_by_board_type):
-        build_attributes = mock.Mock(
-            MbedTargetBuildAttributes, labels={"T"}, features={"F"}, components={"C"}
-        )  # Zero interface safety here, dataclasses don't support spec_set
-        get_build_attributes_by_board_type.return_value = build_attributes
-        mbed_program_directory = Path("some-program")
-        board_type = "K64F"
-
-        subject = BoardLabelFilter(board_type, mbed_program_directory)
-
-        self.assertTrue(subject(Path("TARGET_T")))
-        self.assertTrue(subject(Path("TARGET_T", "FEATURE_F")))
-        self.assertTrue(subject(Path("TARGET_T", "FEATURE_F", "COMPONENT_C")))
-
-        get_build_attributes_by_board_type.assert_called_with(board_type, mbed_program_directory)
-
-
-class TestLabelFilter(TestCase):
-    def test_matches_paths_not_following_label_rules(self):
-        subject = LabelFilter("TARGET", ["BAR", "BAZ"])
-
-        self.assertFalse(subject(Path("mbed-os", "TARGET_FOO", "some_file.c")))
-        self.assertFalse(subject(Path("mbed-os", "TARGET_BAR", "TARGET_FOO", "other_file.c")))
-
-    def test_does_not_match_paths_following_label_rules(self):
-        subject = LabelFilter("TARGET", ["BAR", "BAZ"])
-
-        self.assertTrue(subject(Path("mbed-os", "TARGET_BAR", "some_file.c")))
-        self.assertTrue(subject(Path("mbed-os", "COMPONENT_X", "header.h")))
-        self.assertTrue(subject(Path("mbed-os", "COMPONENT_X", "TARGET_BAZ", "some_file.c")))
-        self.assertTrue(subject(Path("README.md")))
-
-
-class TestMbedignoreFilter(TestCase):
-    def test_matches_files_by_name(self):
-        subject = MbedignoreFilter(("*.py",))
-
-        self.assertFalse(subject("file.py"))
-        self.assertFalse(subject("nested/file.py"))
-        self.assertTrue(subject("file.txt"))
-
-    def test_matches_wildcards(self):
-        subject = MbedignoreFilter(("*/test/*",))
-
-        self.assertFalse(subject("foo/test/bar.txt"))
-        self.assertFalse(subject("bar/test/other/file.py"))
-        self.assertTrue(subject("file.txt"))
-
-    def test_from_file(self):
-        with TemporaryDirectory() as temp_directory:
-            mbedignore = Path(temp_directory, ".mbedignore")
-            mbedignore.write_text(
-                """
-# Comment
-
-foo/*.txt
-*.py
-"""
-            )
-
-            subject = MbedignoreFilter.from_file(mbedignore)
-
-            self.assertEqual(
-                subject._patterns, (str(Path(temp_directory, "foo/*.txt")), str(Path(temp_directory, "*.py")),)
-            )
